Fix bug, update rating_number data type to store 0.5 rating

+++ server/db/schema/04_rating.sql
@@ -4,5 +4,5 @@ CREATE TABLE ratings (
 id SERIAL PRIMARY KEY NOT NULL,
 track_playlist_id INTEGER REFERENCES track_playlists(id) ON DELETE CASCADE,
 user_id VARCHAR(255) NOT NULL, 
-rating_number INTEGER
+rating_number DECIMAL
 );
